+++ morsecode.py
@@ -1,5 +1,4 @@
-# -*- coding: utf8 -*-
-
+import re
 
 # Help Function - 수정하지 말 것
 def get_morse_code_dict():
@@ -9,7 +8,6 @@ def get_morse_code_dict():
         "K": "-.-", "X": "-..-", "J": ".---", "W": ".--", "L": ".-..", "Y": "-.--", "M": "--", "Z": "--.."
     }
     return morse_code
-
 
 # Help Function - 수정하지 말 것
 def get_help_message():
@@ -23,157 +21,33 @@ def get_help_message():
         message += "%s: %s\t" % (key, morse_code[key])
         if counter % 5 == 0:
             message += "\n"
-
     return message
 
-
 def is_help_command(user_input):
-    """
-    Input:
-        - user_input : 문자열값으로 사용자가 입력하는 문자
-    Output:
-        - 입력한 값이 대소문자 구분없이 "H" 또는 "HELP"일 경우 True,
-          그렇지 않을 경우 False를 반환함
-    Examples:
-        >>> import morsecode as mc
-        >>> mc.is_help_command("H")
-        True
-        >>> mc.is_help_command("Help")
-        True
-        >>> mc.is_help_command("Half")
-        False
-        >>> mc.is_help_command("HeLp")
-        True
-        >>> mc.is_help_command("HELLO")
-        False
-        >>> mc.is_help_command("E")
-        False
-    """
-    # ===Modify codes below=============
-    # 조건에 따라 변환되어야 할 결과를 result 변수에 할당 또는 필요에 따라 자유로운 수정
-    result = None
+
+    if user_input.isalpha() and user_input.upper() in ['HELP', 'H']:
+        return True
+    else:
+        return False
+#========================= 인코딩 파트 ========================================
+def is_validated_english_sentence(user_input): #영문장이 유효한 문장인지 체크
+    result = False
+    only_eng = "".join(re.findall('[A-Za-z.,!?]', user_input))
+    if only_eng == user_input.strip().replace(" ",""):
+        if re.sub('[.,!?]','',only_eng) != "": #check empty str
+            result = True
 
     return result
-    # ==================================
 
+def get_cleaned_english_sentence(raw_english_sentence):#유효한 영문장을 대상으로 전처리
 
-def is_validated_english_sentence(user_input):
-    """
-    Input:
-        - user_input : 문자열값으로 사용자가 입력하는 문자
-    Output:
-        - 입력한 값이 아래에 해당될 경우 False, 그렇지 않으면 True
-          1) 숫자가 포함되어 있거나,
-          2) _@#$%^&*()-+=[]{}"';:\|`~ 와 같은 특수문자가 포함되어 있거나
-          3) 문장부호(.,!?)를 제외하면 입력값이 없거나 빈칸만 입력했을 경우
-    Examples:
-        >>> import morsecode as mc
-        >>> mc.is_validated_english_sentence("Hello 123")
-        False
-        >>> mc.is_validated_english_sentence("Hi!")
-        True
-        >>> mc.is_validated_english_sentence(".!.")
-        False
-        >>> mc.is_validated_english_sentence("!.!")
-        False
-        >>> mc.is_validated_english_sentence("kkkkk... ^^;")
-        False
-        >>> mc.is_validated_english_sentence("This is Gachon University.")
-        True
-    """
-    # ===Modify codes below=============
-    # 조건에 따라 변환되어야 할 결과를 result 변수에 할당 또는 필요에 따라 자유로운 수정
-    result = None
+    if re.findall('[.,!?]', raw_english_sentence):
+        raw_english_sentence = re.sub('[.,!?]', '', raw_english_sentence).strip()
 
-    return result
-    # ==================================
-
-
-def is_validated_morse_code(user_input):
-    """
-    Input:
-        - user_input : 문자열값으로 사용자가 입력하는 문자
-    Output:
-        - 입력한 값이 아래에 해당될 경우 False, 그렇지 않으면 True
-          1) "-","."," "외 다른 글자가 포함되어 있는 경우
-          2) get_morse_code_dict 함수에 정의된 Morse Code 부호외 다른 코드가 입력된 경우 ex)......
-    Examples:
-        >>> import morsecode as mc
-        >>> mc.is_validated_morse_code("..")
-        True
-        >>> mc.is_validated_morse_code("..-")
-        True
-        >>> mc.is_validated_morse_code("..-..")
-        False
-        >>> mc.is_validated_morse_code(". . . .")
-        True
-        >>> mc.is_validated_morse_code("-- -- -- --")
-        True
-        >>> mc.is_validated_morse_code("!.1 abc --")
-        False
-    """
-    # ===Modify codes below=============
-    # 조건에 따라 변환되어야 할 결과를 result 변수에 할당 또는 필요에 따라 자유로운 수정
-    result = None
-
-    return result
-    # ==================================
-
-
-
-def get_cleaned_english_sentence(raw_english_sentence):
-    """
-    Input:
-        - raw_english_sentence : 문자열값으로 Morse Code로 변환 가능한 영어 문장
-    Output:
-        - 입력된 영어문장에수 4개의 문장부호를 ".,!?" 삭제하고, 양쪽끝 여백을 제거한 문자열 값 반환
-    Examples:
-        >>> import morsecode as mc
-        >>> mc.get_cleaned_english_sentence("This is Gachon!!")
-        'This is Gachon'
-        >>> mc.get_cleaned_english_sentence("Is this Gachon?")
-        'Is this Gachon'
-        >>> mc.get_cleaned_english_sentence("How are you?")
-        'How are you'
-        >>> mc.get_cleaned_english_sentence("Fine, Thank you. and you?")
-        'Fine Thank you and you'
-    """
-    # ===Modify codes below=============
-    # 조건에 따라 변환되어야 할 결과를 result 변수에 할당 또는 필요에 따라 자유로운 수정
-    result = None
-
-    return result
-    # ==================================
-
-
-def decoding_character(morse_character):
-    """
-    Input:
-        - morse_character : 문자열값으로 get_morse_code_dict 함수로 알파벳으로 치환이 가능한 값의 입력이 보장됨
-    Output:
-        - Morse Code를 알파벳으로 치환함 값
-    Examples:
-        >>> import morsecode as mc
-        >>> mc.decoding_character("-")
-        'T'
-        >>> mc.decoding_character(".")
-        'E'
-        >>> mc.decoding_character(".-")
-        'A'
-        >>> mc.decoding_character("...")
-        'S'
-        >>> mc.decoding_character("....")
-        'H'
-        >>> mc.decoding_character("-.-")
-        'K'
-    """
-    # ===Modify codes below=============
-    # 조건에 따라 변환되어야 할 결과를 result 변수에 할당 또는 필요에 따라 자유로운 수정
-    morse_code_dict = get_morse_code_dict()
-    result = None
-
-    return result
-    # ==================================
+    else:
+        raw_english_sentence = raw_english_sentence.strip()
+    
+    return raw_english_sentence
 
 
 def encoding_character(english_character):
@@ -182,90 +56,84 @@ def encoding_character(english_character):
         - english_character : 문자열값으로 알파벳 한 글자의 입력이 보장됨
     Output:
         - get_morse_code_dict 함수의 반환 값으로 인해 변환된 모스부호 문자열값
-    Examples:
-        >>> import morsecode as mc
-        >>> mc.encoding_character("G")
-        '--.'
-        >>> mc.encoding_character("A")
-        '.-'
-        >>> mc.encoding_character("C")
-        '-.-.'
-        >>> mc.encoding_character("H")
-        '....'
-        >>> mc.encoding_character("O")
-        '---'
-        >>> mc.encoding_character("N")
-        '-.'
     """
-    # ===Modify codes below=============
-    # 조건에 따라 변환되어야 할 결과를 result 변수에 할당 또는 필요에 따라 자유로운 수정
     morse_code_dict = get_morse_code_dict()
-    result = None
-
-    return result
-    # ==================================
-
-
-def decoding_sentence(morse_sentence):
-    """
-    Input:
-        - morse_sentence : 문자열 값으로 모스 부호를 표현하는 문자열
-    Output:
-        - 모스부호를 알파벳으로 변환한 문자열
-    Examples:
-        >>> import morsecode as mc
-        >>> mc.decoding_sentence("... --- ...")
-        'SOS'
-        >>> mc.decoding_sentence("--. .- -.-. .... --- -.")
-        'GACHON'
-        >>> mc.decoding_sentence("..  .-.. --- ...- .  -.-- --- ..-")
-        'I LOVE YOU'
-        >>> mc.decoding_sentence("-.-- --- ..-  .- .-. .  ..-. ")
-        'YOU ARE F'
-    """
-    # ===Modify codes below=============
-    # 조건에 따라 변환되어야 할 결과를 result 변수에 할당 또는 필요에 따라 자유로운 수정
-    result = None
-
-    return result
-    # ==================================
+    return morse_code_dict.get(english_character)
 
 
 def encoding_sentence(english_sentence):
-    """
-    Input:
-        - english_sentence : 문자열 값으로 모스 부호로 변환이 가능한 영어문장
-    Output:
-        - 입력된 영어문장 문자열 값을 모스부호로 변환된 알파벳으로 변환한 문자열
-          단 양쪽 끝에 빈칸은 삭제한다.
-    Examples:
-        >>> import morsecode as mc
-        >>> mc.encoding_sentence("HI! Fine, Thank you.")
-        '.... ..  ..-. .. -. .  - .... .- -. -.-  -.-- --- ..-'
-        >>> mc.encoding_sentence("Hello! This is CS fifty Class.")
-        '.... . .-.. .-.. ---  - .... .. ...  .. ...  -.-. ...  ..-. .. ..-. - -.--  -.-. .-.. .- ... ...'
-        >>> mc.encoding_sentence("We Are Gachon")
-        '.-- .  .- .-. .  --. .- -.-. .... --- -.'
-        >>> mc.encoding_sentence("Hi! Hi!")
-        '.... ..  .... ..'
-    """
-    # ===Modify codes below=============
-    # 조건에 따라 변환되어야 할 결과를 result 변수에 할당 또는 필요에 따라 자유로운 수정
-    result = None
+    morse = get_cleaned_english_sentence(english_sentence).upper()
+    morse = list(morse)
+
+    result = ''
+    for i in range(0, len(morse)):
+        if morse[i] == ' ' and morse[i+1] == ' ':
+            pass
+        elif morse[i] == ' ' and morse[i+1] != ' ':
+            result += ' '
+        else:
+            result += str(encoding_character(morse[i]))
+            result += ' '
+    
+    result = result.strip()
+    return result
+
+#========================== 인코딩 파트 ========================================
+
+#========================== 디코딩 파트 ========================================
+def is_validated_morse_code(user_input):
+
+    codes = user_input.split()
+    reverse_code_dict = dict(map(reversed, get_morse_code_dict().items()))
+
+    for code in codes:
+        if reverse_code_dict.get(code):
+            pass
+        else:
+            return False
+
+    return True
+def decoding_character(morse_character):
+    # 기존 딕셔너리의 key와 value를 reverse 시켜서 값을 빼온다.
+    reverse_code_dict = dict(map(reversed, get_morse_code_dict().items()))
+    return reverse_code_dict.get(morse_character)
+
+def decoding_sentence(morse_sentence):
+    morse = morse_sentence
+    morse = morse.split(' ')
+    
+    result = ''
+    for i in range(len(morse)):
+        if morse[i] == '':
+            result += ' '
+        else:
+            result += str(decoding_character(morse[i]))
 
     return result
-    # ==================================
-
-
+#========================== 디코딩 파트 ========================================
+decoding_sentence("--. --.  --. -  -  -  - . . . .")
 def main():
     print("Morse Code Program!!")
     # ===Modify codes below=============
+    flag = True
+    while flag:
+        user_input = input("Input your message(H- Help, 0- Exit): ")
 
-
-
+        if user_input == '0':
+            break
+        elif is_help_command(user_input):
+            message = get_help_message()
+            print(message)
+        elif is_validated_english_sentence(user_input):
+            result = encoding_sentence(user_input)
+            print(result)
+        elif is_validated_morse_code(user_input):
+            result = decoding_sentence(user_input)
+            print(result)
+        elif user_input == ' ':
+            print("Wrong Input")
+        else:
+            print("Wrong Input")
     # ==================================
     print("Good Bye")
     print("Morse Code Program Finished!!")
-
-if __name__ == "__main__":
-    main()
